+++ src/runtime/qsolverjob.cpp
@@ -119,7 +119,7 @@ void QSolverJob::solving(){
     qDebug().noquote() << tr("Solving done.");//.toStdString() << std::endl;
 }
 
-long QSolverJob::estimate_tree_memory(QString range1,QString range2,QString board){
+long long QSolverJob::estimate_tree_memory(QString range1,QString range2,QString board){
     qDebug().noquote() << tr("Estimating tree memory..");//.toStdString() << endl;
     if(this->mode == Mode::HOLDEM){
         return ps_holdem.estimate_tree_memory(range1,range2,board);
